Values quoted;

+++ plots/root/map/src/wrapper.sh
@@ -1,18 +1,11 @@
 #!/bin/bash
 
 _MBTILES_FILE=data/osm.mbtiles
-_LOCK_FILE=data/download.lock
 
-if [[ -e $_LOCK_FILE ]]; then
-    while [[ -e $_LOCK_FILE ]]; do
-        echo "Waiting for MBtiles..."
-        sleep 3s
-    done
-elif [[ -n "$SERENDITREE_DATA_URL" ]]; then
+if [[ -n "$SERENDITREE_DATA_URL" ]] && [[ ! -e $_MBTILES_FILE ]]; then
     touch $_LOCK_FILE
     echo "Downloading MBtiles..."
     curl -L "$SERENDITREE_DATA_URL" -o $_MBTILES_FILE
-    rm -v $_LOCK_FILE
 fi
 
 if [[ ! -e $_MBTILES_FILE ]]; then
